Reorganize reducer slices for data-explorers store
